Speculative fix for blog delete bug.
Rationale: What looks like is happening is the confirm/cancel was preventing the execution of the delete action, however the remove action overrides this and executes the action call regardless. So, we now check to see if the default behaviour has been overridden (i.e. cancel button pressed) before continuing.

I've not been able to replicate the original fault on my localhost, so perhaps someone else can confirm..?