(#134) Added nginx-based auto-language negotiation for /.
**The Problem:** Upon following the directions to the letter, when I
go to http://localhost:8080/ I get a 403: Forbidden nginx error screen.

**The Reason:** No language detection / negotiation is done, and / is empty.

**The Solution:** I added nginx-based language negotiation using the browser's
`Accept-Language` header. If the user does not set a header (some search engines,
curl, etc.), it will show the English page.

This only applies to the / route. It is assumed that any other route will be in
the desired language.

Fixes #134.