(feature) Indexing email clicks
Extends the minds-metrics-index to track email activity

Emails that build urls with campaign and topic will now be tracked.

Emails that do not will be marked as unknown

Email unsubscribe links will be tracked as email:unsubscribe

This MR requires extending the actiont type in ElasticSearch to use the keyword property. Those are stored in Postman

https://gitlab.com/minds/engine/issues/210