WIP firehose api
New firehose admin endpoints for getting an unsorted list of activites from minds_badger and an admin endpoint that takes an activity guid and will mark it as moderated.

New managers and repositories for talking to elastic search
Redis lock system with some super basic hash keying. These records shouldn't get very big, so we might want to KISS for now. Else, there are distributed locking libraries we should look into

It stores a key in moderation:entity_id:user_id format where the entity is the activity being moderated and user is the admin who "checked it out". These are set to a TTL of 15 minutes for now.
When we query elastic search, we look for the opposite set (any locks that DON'T belong to the user) and for any activity that has not already been moderated.
Requires adding two new fields to badger (in postman)